Refactor code by isolating components
