+++ src/App.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
-import {Layout, Header, Navigation, Drawer, Content, Footer, FooterSection, FooterLinkList, FooterDropDownSection} from 'react-mdl';
+import {Layout, Header, Navigation, Drawer} from 'react-mdl';
+import Content from "./components/Content"
 import {Link} from 'react-router-dom';
-import Main from './components/Main';
 
 class App extends Component {
 
@@ -25,51 +25,7 @@ class App extends Component {
                             <Link to="/question"> Questions </Link>
                         </Navigation>
                     </Drawer>
-                    <Content>
-                        <div className="page-content"/>
-                        <Main/>
-                        <Footer size="mega">
-                            <FooterSection type="middle">
-                                <FooterDropDownSection title="Features">
-                                    <FooterLinkList>
-                                        <a href="/aboutme">About</a>
-                                        <a href="/">Terms</a>
-                                        <a href="/">Partners</a>
-                                        <a href="/">Updates</a>
-                                    </FooterLinkList>
-                                </FooterDropDownSection>
-                                <FooterDropDownSection title="Details">
-                                    <FooterLinkList>
-                                        <a href="/">Specs</a>
-                                        <a href="/">Tools</a>
-                                        <a href="/">Resources</a>
-                                    </FooterLinkList>
-                                </FooterDropDownSection>
-                                <FooterDropDownSection title="Technology">
-                                    <FooterLinkList>
-                                        <a href="/">How it works</a>
-                                        <a href="/">Patterns</a>
-                                        <a href="/">Usage</a>
-                                        <a href="/">Products</a>
-                                        <a href="/">Contracts</a>
-                                    </FooterLinkList>
-                                </FooterDropDownSection>
-                                <FooterDropDownSection title="FAQ">
-                                    <FooterLinkList>
-                                        <a href="/">Questions</a>
-                                        <a href="/">Answers</a>
-                                        <a href="/">Contact Us</a>
-                                    </FooterLinkList>
-                                </FooterDropDownSection>
-                            </FooterSection>
-                            <FooterSection type="bottom" logo="Title">
-                                <FooterLinkList>
-                                    <a href="/">Help</a>
-                                    <a href="/">Privacy & Terms</a>
-                                </FooterLinkList>
-                            </FooterSection>
-                        </Footer>
-                    </Content>
+                    <Content/>
                 </Layout>
             </div>
